+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - update Chinese (Traditional) translation (#590)
 
+### Fixed
+- Enabling/disabling Mini/Long Breaks resets the breaks (#503)
+
 ## [0.99.4] - 2020-05-15
 ### Fixed
 - checkbox across macOS, Windows, Linux